refactor: reorganize component structure; move Button and ProtectedRoute to new directories; update imports accordingly
